+++ spec/dummy/app/components/callback_component.rb
@@ -1,0 +1,7 @@
+class CallbackComponent < ApplicationComponent
+  after_initialize -> { @links = [] }
+
+  def with_link(**options)
+    @links << options
+  end
+end